requirements-dev.txt
-r requirements.txt

black
recommonmark
sphinx
sphinx_rtd_theme
django-reversion
django-autocomplete-light
django-webtest
